+++ interfaces/handler/project.go
@@ -8,7 +8,7 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/traPtitech/traPortfolio/domain"
 	"github.com/traPtitech/traPortfolio/usecases/repository"
-	service "github.com/traPtitech/traPortfolio/usecases/service/project_service"
+	"github.com/traPtitech/traPortfolio/usecases/service"
 )
 
 //TODO 何月？
